+++ CHANGES.rst
@@ -5,6 +5,11 @@ Changelog
 0.2dev
 ------
 
+- Add ``user`` and ``group`` keyword arguments to
+  ``bda.plone.discount.settings.PersistendDiscountSettings.rules`` function.
+  Restrict the use of these keywords by defined ``for_attribute`` scope.
+  [rnix]
+
 - Avoid division / 0 in ``bda.plone.discount.calculator.CartDiscount.vat``
   [rnix]